+++ src/ui/csurf_ui_button_bar.hpp
@@ -21,11 +21,11 @@ static void ReaSonusButtonBar(
     if (ImGui::BeginChild(m_ctx, "button-bar", 0.0, 34.0, ImGui::ChildFlags_FrameStyle))
     {
         ImGui::CalcTextSize(m_ctx, action_label.c_str(), &x_pos_1, &y_pos_1);
-        float buttonWidth1 = x_pos_1 + button_padding_x * 2.f;
+        float buttonWidth1 = (int)x_pos_1 + button_padding_x * 2.f;
         ImGui::CalcTextSize(m_ctx, cancel_label.c_str(), &x_pos_1, &y_pos_1);
-        float buttonWidth2 = x_pos_1 + button_padding_x * 2.f;
+        float buttonWidth2 = (int)x_pos_1 + button_padding_x * 2.f;
         ImGui::GetStyleVar(m_ctx, ImGui::StyleVar_ItemSpacing, &x_pos_2, &y_pos_2);
-        float widthNeeded = buttonWidth1 + x_pos_2 + buttonWidth2;
+        float widthNeeded = (int)buttonWidth1 + x_pos_2 + buttonWidth2;
         ImGui::GetContentRegionAvail(m_ctx, &x_pos_2, &y_pos_2);
 
         ImGui::SetCursorPosX(m_ctx, ImGui::GetCursorPosX(m_ctx) + x_pos_2 - widthNeeded - 4);
